+++ src/main/java/org/spin/pos/service/bank/BankManagement.java
@@ -15,9 +15,28 @@
 
 package org.spin.pos.service.bank;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import org.adempiere.core.domains.models.I_C_Bank;
+import org.adempiere.core.domains.models.I_C_BankAccount;
 import org.adempiere.exceptions.AdempiereException;
 import org.compiere.model.MBank;
+import org.compiere.model.MRole;
+import org.compiere.model.Query;
 import org.compiere.util.Env;
+import org.compiere.util.Util;
+import org.spin.backend.grpc.common.BankAccount;
+import org.spin.backend.grpc.pos.Bank;
+import org.spin.backend.grpc.pos.ListBankAccountsRequest;
+import org.spin.backend.grpc.pos.ListBankAccountsResponse;
+import org.spin.backend.grpc.pos.ListBanksRequest;
+import org.spin.backend.grpc.pos.ListBanksResponse;
+import org.spin.base.db.LimitUtil;
+import org.spin.base.util.ConvertUtil;
+import org.spin.base.util.SessionManager;
+import org.spin.base.util.ValueUtil;
+import org.spin.pos.util.POSConvertUtil;
 
 /**
  * A util class for change values for documents
@@ -34,6 +53,92 @@ public class BankManagement {
 			throw new AdempiereException("@C_Bank_ID@ @NotFound@");
 		}
 		return bank;
+	}
+	
+	public static ListBanksResponse.Builder listBanks(ListBanksRequest request) {
+		List<Object> filtersList = new ArrayList<>();
+
+		String whereClause = "BankType = 'B' ";
+		if (!Util.isEmpty(request.getSearchValue(), false)) {
+			filtersList.add(request.getSearchValue());
+			whereClause += "AND UPPER(Name) LIKE '%' || UPPER(?) || '%' ";
+		}
+
+		Query query = new Query(
+			Env.getCtx(),
+			I_C_Bank.Table_Name,
+			whereClause,
+			null
+		)
+			.setParameters(filtersList)
+			.setOnlyActiveRecords(true)
+			.setApplyAccessFilter(MRole.SQL_FULLYQUALIFIED, MRole.SQL_RO)
+		;
+
+		//	Get page and count
+		String nexPageToken = null;
+		int pageNumber = LimitUtil.getPageNumber(SessionManager.getSessionUuid(), request.getPageToken());
+		int limit = LimitUtil.getPageSize(request.getPageSize());
+		int offset = (pageNumber - 1) * limit;
+		int recordCount = query.count();
+
+		ListBanksResponse.Builder builderList = ListBanksResponse.newBuilder()
+			.setRecordCount(recordCount)
+			.setNextPageToken(
+				ValueUtil.validateNull(nexPageToken)
+			)
+		;
+
+		//	Get List
+		List<Integer> banksIdsList = query
+			.setLimit(limit, offset)
+			.getIDsAsList()
+		;
+
+		banksIdsList.forEach(tableId -> {
+			Bank.Builder accountingDocument = POSConvertUtil.convertBank(tableId);
+			builderList.addRecords(accountingDocument);
+		});
+
+		return builderList;
+	}
+	
+	public static ListBankAccountsResponse.Builder listBankAccounts(ListBankAccountsRequest request) {
+		// validate and get Bank
+		MBank bank = BankManagement.validateAndGetBank(request.getBankId());
+
+		List<Object> filtersList = new ArrayList<>();
+		filtersList.add(bank.getC_Bank_ID());
+		String whereClause = "C_Bank_ID = ? ";
+		if (!Util.isEmpty(request.getSearchValue(), false)) {
+			filtersList.add(request.getSearchValue());
+			whereClause = "AND UPPER(Name) LIKE '%' || UPPER(?) || '%' ";
+		}
+
+		Query query = new Query(
+			Env.getCtx(),
+			I_C_BankAccount.Table_Name,
+			whereClause,
+			null
+		)
+			.setParameters(filtersList)
+			.setOnlyActiveRecords(true)
+			.setApplyAccessFilter(MRole.SQL_FULLYQUALIFIED, MRole.SQL_RO)
+		;
+
+		//	Get page and count
+		int recordCount = query.count();
+
+		ListBankAccountsResponse.Builder builderList = ListBankAccountsResponse.newBuilder()
+			.setRecordCount(recordCount)
+		;
+
+		//	Get List
+		query.getIDsAsList().forEach(bankAccountId -> {
+			BankAccount.Builder bankAccountBuilder = ConvertUtil.convertBankAccount(bankAccountId);
+			builderList.addRecords(bankAccountBuilder);
+		});
+		return builderList;
 	}
 
 }
