+++ src/main/proto/point_of_sales.proto
@@ -1235,6 +1235,8 @@ message Payment {
 	Charge charge = 24;
 	DocumentType document_type = 25;
 	bool is_processed = 27;
+	string order_document_no = 28;
+	string invoice_document_no = 29;
 }
 
 // Request for delete a order
@@ -1369,6 +1371,7 @@ message OrderLine {
 	ProductConversion product_uom = 31;
 	time_control.ResourceAssignment resource_assignment = 32;
 	int32 id = 33;
+	int32 source_rma_line_id = 34;
 }
 
 // Request Product Price
